requirements.txt
gpt-researcher
pandas
tavily-python
python-dotenv
langchain
streamlit